# temp_openapi.json
{"openapi":"3.1.0","info":{"title":"InsightReader API","description":"上下文智能探针 API - V2.0 版本","version":"1.0.0-mvp"},"paths":{"/api/v1/insights/generate":{"post":{"tags":["insights"],"summary":"Generate Insight","description":"流式生成洞察卡片\n\nArgs:\n    request: 包含选中文本、上下文和意图的请求\n\nReturns:\n    SSE 流式响应","operationId":"generate_insight_api_v1_insights_generate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InsightRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/insights/health":{"get":{"tags":["insights"],"summary":"Health","description":"API 健康检查","operationId":"health_api_v1_insights_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/auth/google/login":{"get":{"tags":["auth"],"summary":"Google Login","description":"Google OAuth 登录 - 重定向到 Google","operationId":"google_login_api_v1_auth_google_login_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/auth/google/callback":{"get":{"tags":["auth"],"summary":"Google Callback","description":"Google OAuth 回调","operationId":"google_callback_api_v1_auth_google_callback_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/auth/github/login":{"get":{"tags":["auth"],"summary":"Github Login","description":"GitHub OAuth 登录 - 重定向到 GitHub","operationId":"github_login_api_v1_auth_github_login_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/auth/github/callback":{"get":{"tags":["auth"],"summary":"Github Callback","description":"GitHub OAuth 回调","operationId":"github_callback_api_v1_auth_github_callback_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/auth/magic-link/request":{"post":{"tags":["auth"],"summary":"Request Magic Link","description":"请求魔法链接","operationId":"request_magic_link_api_v1_auth_magic_link_request_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MagicLinkRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/magic-link/verify":{"get":{"tags":["auth"],"summary":"Verify Magic Link Endpoint","description":"验证魔法链接并登录","operationId":"verify_magic_link_endpoint_api_v1_auth_magic_link_verify_get","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string","title":"Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["auth"],"summary":"Get Current User Info","description":"获取当前用户信息","operationId":"get_current_user_info_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/auth/logout":{"post":{"tags":["auth"],"summary":"Logout","description":"登出（前端删除 token 即可）","operationId":"logout_api_v1_auth_logout_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/collections/":{"post":{"tags":["collections"],"summary":"Create Insight Card","description":"创建洞察卡片（收藏）","operationId":"create_insight_card_api_v1_collections__post","security":[{"OAuth2PasswordBearer":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InsightCardCreate"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InsightCardResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["collections"],"summary":"Get Insight Cards","description":"获取用户的洞察卡片列表","operationId":"get_insight_cards_api_v1_collections__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"default":20,"title":"Limit"}},{"name":"intent","in":"query","required":false,"schema":{"anyOf":[{"type":"string","pattern":"^(explain|analyze|counter)$"},{"type":"null"}],"title":"Intent"}},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Search"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InsightCardListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/collections/{card_id}":{"get":{"tags":["collections"],"summary":"Get Insight Card","description":"获取单个洞察卡片","operationId":"get_insight_card_api_v1_collections__card_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"card_id","in":"path","required":true,"schema":{"type":"integer","title":"Card Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InsightCardResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["collections"],"summary":"Update Insight Card","description":"更新洞察卡片（主要用于更新标签）","operationId":"update_insight_card_api_v1_collections__card_id__patch","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"card_id","in":"path","required":true,"schema":{"type":"integer","title":"Card Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InsightCardUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InsightCardResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["collections"],"summary":"Delete Insight Card","description":"删除洞察卡片","operationId":"delete_insight_card_api_v1_collections__card_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"card_id","in":"path","required":true,"schema":{"type":"integer","title":"Card Id"}}],"responses":{"204":{"description":"Successful Response"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","description":"根路由","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health":{"get":{"summary":"Health Check","description":"健康检查","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"InsightCardCreate":{"properties":{"article_title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Article Title"},"article_content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Article Content"},"selected_text":{"type":"string","title":"Selected Text"},"context":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Context"},"intent":{"type":"string","pattern":"^(explain|analyze|counter)$","title":"Intent"},"custom_question":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Custom Question"},"insight":{"type":"string","title":"Insight"},"model_used":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model Used"},"tokens":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Tokens"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags"}},"type":"object","required":["selected_text","intent","insight"],"title":"InsightCardCreate","description":"创建洞察卡片请求"},"InsightCardListResponse":{"properties":{"total":{"type":"integer","title":"Total"},"items":{"items":{"$ref":"#/components/schemas/InsightCardResponse"},"type":"array","title":"Items"}},"type":"object","required":["total","items"],"title":"InsightCardListResponse","description":"洞察卡片列表响应"},"InsightCardResponse":{"properties":{"id":{"type":"integer","title":"Id"},"user_id":{"type":"integer","title":"User Id"},"article_title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Article Title"},"selected_text":{"type":"string","title":"Selected Text"},"context":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Context"},"intent":{"type":"string","title":"Intent"},"custom_question":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Custom Question"},"insight":{"type":"string","title":"Insight"},"model_used":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Model Used"},"tokens":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Tokens"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags"}},"type":"object","required":["id","user_id","article_title","selected_text","context","intent","custom_question","insight","model_used","tokens","created_at","tags"],"title":"InsightCardResponse","description":"洞察卡片响应"},"InsightCardUpdate":{"properties":{"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags"}},"type":"object","title":"InsightCardUpdate","description":"更新洞察卡片请求"},"InsightRequest":{"properties":{"selected_text":{"type":"string","maxLength":500,"minLength":1,"title":"Selected Text","description":"用户选中的文本"},"context":{"type":"string","maxLength":2000,"minLength":10,"title":"Context","description":"选中文本的上下文"},"intent":{"type":"string","pattern":"^(explain|analyze|counter)$","title":"Intent","description":"用户意图: explain(解释) | analyze(分析) | counter(反驳)"},"custom_question":{"anyOf":[{"type":"string","maxLength":200},{"type":"null"}],"title":"Custom Question","description":"用户自定义问题（V2.0）"}},"type":"object","required":["selected_text","context","intent"],"title":"InsightRequest","description":"洞察生成请求","examples":[{"context":"在伦理学中，康德的绝对命令是一个核心概念。它要求我们的行为准则能够成为普遍法则。","intent":"explain","selected_text":"康德的绝对命令"}]},"MagicLinkRequest":{"properties":{"email":{"type":"string","format":"email","title":"Email"}},"type":"object","required":["email"],"title":"MagicLinkRequest","description":"魔法链接请求"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"email":{"type":"string","title":"Email"},"username":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Username"},"avatar_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Avatar Url"},"oauth_provider":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Oauth Provider"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"last_login":{"type":"string","format":"date-time","title":"Last Login"},"is_active":{"type":"boolean","title":"Is Active"}},"type":"object","required":["id","email","created_at","last_login","is_active"],"title":"UserResponse","description":"用户响应"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/auth/login"}}}}}}